Added new Search methods
In this commit new way of filter contacts has been added. Now for searching contacts you can use not only name but number as well. And even if you type number without parenthesis or dashes, or if you type with parenthesis and with spaces, but in search field without them, in both cases it will work.